# gateway/gateway-controller/api/management-openapi.yaml
openapi: 3.0.3
info:
  title: Gateway Controller Management API
  description: |
    REST API for managing API configurations in the WSO2 API Platform Gateway.
  version: 1.0.0
  contact:
    name: WSO2 API Platform Team
servers:
  - url: http://localhost:9090/api/management/v1
    description: Local development
  - url: http://gateway-controller:9090/api/management/v1
    description: Docker/Kubernetes deployment
security:
  - basicAuth: []
paths:
...
  /certificates:
    get:
      summary: List all custom certificates
      description: |
        Retrieve custom TLS certificates currently loaded in the certificate store.
        With usage: upstream (or no usage filter), these are used for verifying
        HTTPS upstream connections. With usage: client, these are the pooled
        client certificate authorities used to authenticate mutual-TLS callers.
      operationId: listCertificates
      x-basicauth-roles: [admin, developer]
      tags:
        - Certificate Management
      parameters:
        - name: usage
          in: query
          required: false
          description: Filter the list to a single usage. Omit to list all certificates regardless of usage.
          schema:
            type: string
            enum: [ upstream, client ]
      responses:
        "200":
          description: List of certificates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificateListResponse"
        "400":
          description: Invalid usage filter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      summary: Upload a new certificate
      description: |
        Upload a new TLS certificate (PEM format) to the Gateway. The certificate is
        loaded dynamically without restarting the Gateway. With usage: upstream (the
        default), it is used for verifying HTTPS upstream connections. With usage:
        client, it becomes a pooled client certificate authority used to authenticate
        mutual-TLS callers; the two purposes never share a trust bundle.
      operationId: uploadCertificate
      x-basicauth-roles: [admin]
      tags:
        - Certificate Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificateUploadRequest"
            example:
              name: my-custom-ca
              certificate: |
                -----BEGIN CERTIFICATE-----
                MIIDXTCCAkWgAwIBAgIJAKL0UG+mRKtjMA0GCSqGSIb3DQEBCwUAMEUxCzAJBgNV
                ...
                -----END CERTIFICATE-----
      responses:
        "201":
          description: Certificate uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificateResponse"
        "400":
          description: Invalid certificate format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: A certificate with this name already exists (names are one namespace across usages)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Request body exceeds the maximum allowed size
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          description: High-level error description
          example: Configuration validation failed
        errors:
          type: array
          description: Detailed validation errors
          items:
            $ref: "#/components/schemas/ValidationError"
    ValidationError:
      type: object
      properties:
        field:
          type: string
          description: Field that failed validation
          example: spec.context
        message:
          type: string
          description: Human-readable error message
          example: Context must start with / and cannot end with /
...
    CertificateUploadRequest:
      type: object
      required:
        - certificate
        - name
      properties:
        name:
          type: string
          description: Unique name for the certificate. Must be unique across all certificates, regardless of usage.
          pattern: "^[a-zA-Z0-9._-]+$"
          minLength: 1
          maxLength: 100
          example: my-custom-ca
        certificate:
          type: string
          description: PEM-encoded X.509 certificate(s). Can contain multiple certificates.
          example: |
            -----BEGIN CERTIFICATE-----
            MIIDXTCCAkWgAwIBAgIJAKL0UG+mRKtjMA0GCSqGSIb3DQEBCwUAMEUxCzAJBgNV
            ...
            -----END CERTIFICATE-----
        usage:
          type: string
          description: |
            Whether this certificate is used to verify upstream/backend HTTPS
            connections (the default), pooled as a client certificate
            authority used to authenticate mutual-TLS callers, or (identity)
            a gateway identity — a certificate chain (leaf first) plus its
            private key that the gateway presents to a backend requiring
            mutual TLS on outbound connections.
          enum: [ upstream, client, identity ]
          default: upstream
          example: upstream
        role:
          type: string
          description: |
            Only meaningful when usage is client. "client" (the default)
            means the authority validates a client certificate presented
            directly on the mTLS connection; "relay" means it validates a
            client certificate relayed via a header (e.g. from a
            terminating load balancer) instead.
          enum: [ client, relay ]
          example: client
        match:
          allOf:
            - $ref: "#/components/schemas/CertificateMatch"
          description: |
            Only valid for role: relay. Narrows which connections
            authenticated as this relay entry can make a relayed header
            believed. Omit to accept any connection presenting a certificate
            from this authority, with no further narrowing.
        privateKey:
          type: string
          writeOnly: true
          description: >
            Required (and only valid) when usage is identity: a PEM-encoded,
            unencrypted private key matching the leaf certificate
            (RSA/ECDSA/Ed25519, PKCS#8/PKCS#1/SEC1). Encrypted at rest by
            the gateway; never returned by any response.
          example: |
            -----BEGIN PRIVATE KEY-----
            MIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQC...
            -----END PRIVATE KEY-----
    CertificateMatch:
      type: object
      description: |
        Narrows a role: relay pool entry to the connections it vouches for.
        Each list, when present, must name at least one non-empty SAN.
      properties:
        dnsSANs:
          type: array
          items:
            type: string
          example: [ "lb.corp.test" ]
        uriSANs:
          type: array
          items:
            type: string
          example: [ "urn:partner-a:edge-lb" ]
    CertificateResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier (UUID) for the certificate
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          description: Name of the certificate
          example: my-custom-ca
        subject:
          type: string
          description: Certificate subject DN (identity certificate, for a client-CA bundle)
          example: CN=My CA,O=My Organization,C=US
        issuer:
          type: string
          description: Certificate issuer DN (identity certificate, for a client-CA bundle)
          example: CN=My CA,O=My Organization,C=US
        notAfter:
          type: string
          description: Certificate expiration date (identity certificate, for a client-CA bundle)
          format: date-time
          example: "2026-11-26 06:07:26"
        count:
          type: integer
          description: Number of certificates in the file
          example: 1
        usage:
          type: string
          description: Whether this is upstream/backend trust, a pooled client certificate authority, or a gateway identity.
          enum: [ upstream, client, identity ]
          example: upstream
        role:
          type: string
          description: Only present for usage client. Whether the authority validates a directly-presented or header-relayed client certificate.
          enum: [ client, relay ]
          example: client
        match:
          allOf:
            - $ref: "#/components/schemas/CertificateMatch"
          description: Only present for role relay entries that were stored with a narrowing match.
        isLeaf:
          type: boolean
          description: True when the identity certificate is not itself a certificate authority (pooled as a one-member authority).
          example: false
        keyAlgorithm:
          type: string
          description: Only present for usage identity. The leaf private key's algorithm (RSA, ECDSA or Ed25519).
          example: RSA
        chainLength:
          type: integer
          description: Only present for usage identity. Number of certificates in the uploaded chain.
          example: 1
        warnings:
          type: array
          description: Non-fatal findings about the upload (e.g. leaf certificate, not yet valid). Omitted when there are none.
          items:
            $ref: "#/components/schemas/CertificateWarning"
        referencedByApis:
          type: integer
          description: Number of deployed APIs referencing this certificate — a client authority (accept/pool) or a gateway identity (upstreamDefinitions[].tls.identity). Only present for usage client or identity.
          example: 0
        message:
          type: string
          description: Success or informational message
          example: Certificate uploaded and SDS updated successfully
        status:
          type: string
          enum: [ success, error ]
          example: success
    CertificateWarning:
      type: object
      properties:
        code:
          type: string
          enum:
            - CLIENT_CA_IS_LEAF
            - CLIENT_CA_NOT_YET_VALID
            - CERT_EXPIRES_SOON
            - IDENTITY_NO_CLIENTAUTH_EKU
          example: CLIENT_CA_IS_LEAF
        field:
          type: string
          example: certificate
        message:
          type: string
          example: the certificate is not a certificate authority; it is pooled as a one-member authority that trusts exactly this certificate
    CertificateListResponse:
      type: object
      properties:
        certificates:
          type: array
          items:
            $ref: "#/components/schemas/CertificateResponse"
...